requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
coverage==7.0.5
decorator==5.1.1
distlib==0.3.8
Django==4.2.7
django-cors-headers==4.3.1
django-crontab==0.7.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
filelock==3.13.4
flake8==7.0.0
google-api-core==2.11.0
google-api-python-client==2.101.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.0
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.2.0
pluggy==1.0.0
prompt-toolkit==3.0.41
protobuf==4.22.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.0.9
pytest==7.2.1
pytest-cov==4.0.0
python-decouple==3.8
python-dotenv==1.0.1
pytils==0.4.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.25.3
wcwidth==0.2.12
